backend/requirements.txt
Flask
gunicorn
Flask-CORS
google-generativeai
google.ai.generativelanguage
google-cloud-firestore
firebase-admin